requirements.txt
python-dotenv==0.16.0
matplotlib==3.4.0
networkx==2.5
numpy==1.20.2
